requirements.txt
pandas
numpy
scikit-learn
openai
streamlit
matplotlib
seaborn
python-dotenv
langchain
langchain-openai
langchain-community